backend/requirements.txt
fastapi
uvicorn
python-multipart
rembg
mediapipe
opencv-python
numpy
Pillow
pandas
scikit-learn
torch
torchvision
